+++ packages/frameworks/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @vercel/frameworks
 
+## 3.0.1
+
+### Patch Changes
+
+- Swap jest for vitest in CLI unit tests ([#11302](https://github.com/vercel/vercel/pull/11302))
+
+## 3.0.0
+
+### Major Changes
+
+- Make "remix" framework preset supersede "vite" ([#11031](https://github.com/vercel/vercel/pull/11031))
+
 ## 2.0.6
 
 ### Patch Changes